+++ src/snowflake/snowpark/stored_procedure_profiler.py
@@ -7,17 +7,17 @@ from typing import List, Literal, Optional
 import snowflake.snowpark
 from snowflake.snowpark._internal.utils import (
     SNOWFLAKE_ANONYMOUS_CALL_WITH_PATTERN,
-    validate_object_name,
+    parse_table_name,
 )
 
 
 class StoredProcedureProfiler:
     """
-    Set up profiler to receive profiles of stored procedures.
+    Set up profiler to receive profiles of stored procedures. This feature cannot be used in owner's right stored
+    procedure because owner's right stored procedure will not be able to set session-level parameters.
 
-    Note:
-        This feature cannot be used in owner's right stored procedure because owner's right stored procedure will not be
-        able to set session-level parameters.
+    See more details about stored procedure profiler at:
+    https://docs.snowflake.com/LIMITEDACCESS/stored-procedures-python-profiler
     """
 
     def __init__(
@@ -29,41 +29,34 @@ class StoredProcedureProfiler:
         self._lock = threading.RLock()
         self._active_profiler_number = 0
 
-    def register_modules(self, stored_procedures: List[str]) -> None:
+    def register_modules(self, stored_procedures: Optional[List[str]] = None) -> None:
         """
         Register stored procedures to generate profiles for them.
 
-        Note:
-            Registered modules will be overwritten by this function. Use this function with an empty string will remove
-            registered modules.
         Args:
-            stored_procedures: List of names of stored procedures.
+            stored_procedures: List of names of stored procedures. Registered modules will be overwritten by this input.
+            Input None or an empty list will remove registered modules.
         """
-        sql_statement = (
-            f"alter session set python_profiler_modules='{','.join(stored_procedures)}'"
-        )
+        sp_string = ",".join(stored_procedures) if stored_procedures is not None else ""
+        sql_statement = f"alter session set python_profiler_modules='{sp_string}'"
         self._session.sql(sql_statement)._internal_collect_with_tag_no_telemetry()
 
-    def set_targeted_stage(self, stage: str) -> None:
+    def set_target_stage(self, stage: str) -> None:
         """
         Set targeted stage for profiler output.
-
-        Note:
-            The stage name must be a fully qualified name.
 
         Args:
             stage: String of fully qualified name of targeted stage
         """
-        validate_object_name(stage)
-        db_name, schema_name, stage_name = stage.split(".")
+        names = parse_table_name(stage)
+        if len(names) != 3:
+            raise ValueError(
+                f"stage name must be fully qualified name, got {stage} instead"
+            )
         existing_stages = self._session.sql(
-            f"show stages like '{stage_name}' in database {db_name}"
+            f"show stages like '{names[2]}' in schema {names[0]}.{names[1]}"
         )._internal_collect_with_tag_no_telemetry()
-        existing_stages_schema_list = [row.schema_name for row in existing_stages]  # type: ignore
-        if (
-            existing_stages_schema_list == []
-            or schema_name not in existing_stages_schema_list
-        ):
+        if len(existing_stages) == 0:
             self._session.sql(
                 f"create temp stage if not exists {stage} FILE_FORMAT = (RECORD_DELIMITER = NONE FIELD_DELIMITER = NONE )"
             )._internal_collect_with_tag_no_telemetry()
@@ -71,16 +64,14 @@ class StoredProcedureProfiler:
         self._session.sql(sql_statement)._internal_collect_with_tag_no_telemetry()
 
     def set_active_profiler(
-        self, active_profiler_type: Literal["LINE", "MEMORY"]
+        self, active_profiler_type: Literal["LINE", "MEMORY"] = "LINE"
     ) -> None:
         """
         Set active profiler.
 
-        Note:
-            Active profiler must be either 'LINE' or 'MEMORY' (case-sensitive). Active profiler is 'LINE' by default.
         Args:
             active_profiler_type: String that represent active_profiler, must be either 'LINE' or 'MEMORY'
-            (case-sensitive).
+            (case-sensitive). Active profiler is 'LINE' by default.
 
         """
         with self._lock:
@@ -128,10 +119,10 @@ class StoredProcedureProfiler:
 
     def get_output(self) -> str:
         """
-        Return the profiles of last executed stored procedure in current thread.
+        Return the profiles of last executed stored procedure in current thread. If there is no previous
+        stored procedure call, an error will be raised.
+        Please call this function right after the stored procedure you want to profile to avoid any error.
 
-        Note:
-            This function must be called right after the execution of stored procedure you want to profile.
         """
         query_id = self._get_last_query_id()
         if query_id is None: